backend/hardwareManager/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
python-dotenv==1.0.0
SQLAlchemy==2.0.21
alembic==1.12.0
fastapi-mail
python-multipart==0.0.6
psycopg2

# Core dependencies
opencv-python>=4.5.0  # Add OpenCV for camera operations
pypylon>=2.2.0

pywin32==306
# FastAPI related dependencies
pydantic>=2.0.0

# Additional utilities that might be needed
aiofiles>=0.8.0
asyncio>=3.4.3